% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendReportMAGICC7.R
\name{appendReportMAGICC7}
\alias{appendReportMAGICC7}
\title{appendReportMAGICC7}
\usage{
appendReportMAGICC7(resultsPath, scenario, dir = ".")
}
\arguments{
\item{resultsPath}{full file path of the excel with the for MAGICC7 results}

\item{scenario}{name of the scenario}

\item{dir}{scenario output directory}
}
\description{
Reads in the results from the MAGICC7 AR6 probabilistic scenarios
}
\author{
Michael Crawford
}
