% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendReportNutrientSurplus.R
\name{appendReportNutrientSurplus}
\alias{appendReportNutrientSurplus}
\title{appendReportNutrientSurplus}
\usage{
appendReportNutrientSurplus(scenario, dir = ".")
}
\arguments{
\item{scenario}{name of the scenario}

\item{dir}{scenario output directory}
}
\description{
Calculates the global nutrient surplus from the grid-level dataset,
nutrientSurplus_total.mz. This must be calculated from the grid-level data because the
nutrient surplus for nonAg land is currently only available at the grid level.
}
\author{
Michael Crawford
}
