% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignScenarioColors.R
\name{assignScenarioColors}
\alias{assignScenarioColors}
\title{validationFSDP}
\usage{
assignScenarioColors(scenario_names)
}
\arguments{
\item{scenario_names}{scenario names to be checked for their colorcode.}
}
\description{
Validation for FSDP MAgPIE runs
written by openAI chatbot.
The command was: "Please write a function in R. It should receive a vector of scenario_names.
In the function, there is a mapping that assigns each element of potential_names to an individual color.
The function checks which elements of scenario_names are in potential_names and returns the
respective colors as a vector of colors.
}
\author{
openAI, Benjamin Bodirsky
}
