% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignScenarioLinetype.R
\name{assignScenarioLinetype}
\alias{assignScenarioLinetype}
\title{assignScenarioLinetype}
\usage{
assignScenarioLinetype(scenario_names)
}
\arguments{
\item{scenario_names}{scenario names to be checked for their colorcode.}
}
\description{
Line Type for FSDP MAgPIE runs
}
\author{
Florian Humpenoeder
}
