% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundlesFSDP.R
\name{bundlesFSDP}
\alias{bundlesFSDP}
\title{bundlesFSDP}
\usage{
bundlesFSDP(repReg, regionSel = "GLO", file = NULL)
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{regionSel}{Region that should be plotted}

\item{file}{file name (e.g. BundleFSDP.pdf or BundleFSDP.jpg) or NULL}
}
\value{
depends on file argument. If file==NULL, the default, the plot is returned.
If a filename is provided the plot is saved as file
}
\description{
FDSP plot with single measure and bundles
}
\details{
plot with single measure and bundles
}
\author{
Florian Humpenoeder
}
