% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFSDP.R
\name{dashboardFSDP}
\alias{dashboardFSDP}
\title{dashboardFSDP}
\usage{
dashboardFSDP(repReg, regionSel = "GLO", outputDir = "")
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{regionSel}{Region that should be plotted}

\item{outputDir}{output directory}
}
\value{
html file
}
\description{
FDSP dashboard
}
\details{
FDSP dashboard
}
\author{
Florian Humpenoeder
}
