% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateHealthImpacts.R
\name{integrateHealthImpacts}
\alias{integrateHealthImpacts}
\title{integrateHealthImpacts}
\usage{
integrateHealthImpacts(healthImpacts_path, scenario, dir = ".")
}
\arguments{
\item{healthImpacts_path}{full pathname of the health impacts dataset}

\item{scenario}{name of the scenario}

\item{dir}{scenario output directory}
}
\description{
Reads in the results from Marco Springmann's health impacts model, filters them down to the
FSEC-relevant variables, and saves them to the selected scenarios' report.mif.
}
\author{
Michael Crawford
}
