% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplotFSDP.R
\name{lineplotFSDP}
\alias{lineplotFSDP}
\title{lineplotFSDP}
\usage{
lineplotFSDP(repReg, val, regionSel = "GLO", file = NULL, scens = "bundles")
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{val}{rds file or data.frame with validation data}

\item{regionSel}{Region that should be plotted (e.g. c("IND","EUR","GLO")). Aggregate will return LIR, MIR and HIR.}

\item{file}{file name}

\item{scens}{if "BAU_FSEC", BAU and FSEC scenarios are plotted, with "central" plots the core scenarios, "extended" plot the core scenarios and all ssps.}
}
\description{
Line plots FSDP
}
\details{
creates validation for FSDP MAgPIE runs
}
\author{
Florian Humpenoeder
}
