% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4fsdp-package.R
\docType{package}
\name{m4fsdp-package}
\alias{m4fsdp}
\alias{m4fsdp-package}
\title{m4fsdp: MAgPIE outputs R package for MAgPIE version 4.x to create outputs for FSDP project}
\description{
Output routines for extracting results from the MAgPIE framework (versions 4.x) for the FSDP project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/m4fsdp}
  \item \url{https://doi.org/10.5281/zenodo.7899913}
  \item Report bugs at \url{https://github.com/pik-piam/m4fsdp/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Leon Bodirsky \email{bodirsky@pik-potsdam.de}

Authors:
\itemize{
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}
  \item Florian Humpenoeder \email{humpenoeder@pik-potsdam.de}
}

}
