% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestoneTable.R
\name{milestoneTable}
\alias{milestoneTable}
\title{milestoneTable}
\usage{
milestoneTable(scenarioFolder, outFolder = NULL, file = NULL)
}
\arguments{
\item{scenarioFolder}{outputfolder of a specific MAgPIE run (should be an FSECe_FSDP run for SI table)}

\item{outFolder}{main outputfolder, where output csv will be saved}

\item{file}{name to save table as csv}
}
\description{
Creates csv with values for SI table on milestones
}
