% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFSDP.R
\name{plotFSDP}
\alias{plotFSDP}
\title{plotFSDP}
\usage{
plotFSDP(
  outputfolder = "output",
  reg = NULL,
  iso = NULL,
  grid = NULL,
  val = NULL,
  reg2iso = NULL,
  rev = NULL,
  dirFsdp = NULL
)
}
\arguments{
\item{outputfolder}{output folder}

\item{reg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script. NULL will automatically detect the most recent version.}

\item{iso}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script. NULL will automatically detect the most recent version.}

\item{grid}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script. NULL will automatically detect the most recent version.#' @details uses the most recent vXX_reg/grid/iso.rds files in the "output" folder by default}

\item{val}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script. NULL will automatically detect the most recent version.#' @return plots in "output" folder}

\item{reg2iso}{rds file or data.frame with Mapping, produced with FSDP_collect.R output script. NULL will automatically detect the most recent version.#' @author Florian Humpenoeder}

\item{rev}{revision (e.g. "v28"). NULL will automatically detect the most recent version.}

\item{dirFsdp}{path to outputfolder of the FSDP run (to create the milestones table). NULL will automatically detect the most recent version}
}
\description{
Creates all FSDP plots
}
