% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialMapsFSDP.R
\name{spatialMapsFSDP}
\alias{spatialMapsFSDP}
\title{spatialMapsFSDP}
\usage{
spatialMapsFSDP(repReg, repIso, repGrid, reg2iso, file = NULL)
}
\arguments{
\item{repReg}{reporting .rds file or data.frame with regional results (produced by FDSP_collect.R output script)}

\item{repIso}{reporting .rds file or data.frame with country level results (produced by FDSP_collect.R output script)}

\item{repGrid}{reporting .rds file or data.frame with grid level results (produced by FDSP_collect.R output script)}

\item{reg2iso}{mapping file or data.frame with regions and countries (produced by FDSP_collect.R output script)}

\item{file}{file name (e.g. FSDP_spatialmaps.pdf or FSDP_spatialmaps.pdf) or NULL}
}
\value{
if file is NULL a ggplot2 object will be return
}
\description{
creates a spatial mapsfor FSDP MAgPIE runs
}
\details{
blub
}
\author{
Florian Humpenoeder
}
