% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplPlotsCropShr.R
\name{SupplPlotsCropShr}
\alias{SupplPlotsCropShr}
\title{SupplPlotsCropShr}
\usage{
SupplPlotsCropShr(gdx, png = NULL)
}
\arguments{
\item{gdx}{gdx File}

\item{png}{Name of png file output in the form *.png}
}
\value{
Crop share on the y-axis and cropland area in each cluster on the x-axis.
}
\description{
Creates crop share plot
}
\details{
blub
}
\author{
Patrick v. Jeetze
}
