% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplPlotsFSDP.R
\name{supplPlotsFSDP}
\alias{supplPlotsFSDP}
\title{supplPlotsFSDP}
\usage{
supplPlotsFSDP(repReg, scenarioType = "all", file = NULL)
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{scenarioType}{options: all, or one of the a-e groupings}

\item{file}{file name to save to}
}
\value{
if file is NULL a ggplot2 object will be return
}
\description{
creates supplementary plots for FSDP MAgPIE runs
}
\details{
blub
}
\author{
David M Chen
}
