% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplPlotsIndia.R
\name{SupplPlotsFSDPIndia}
\alias{SupplPlotsFSDPIndia}
\title{SupplPlotsFSDPIndia}
\usage{
SupplPlotsFSDPIndia(
  repReg,
  scenarioType = "all",
  save = TRUE,
  outputdir = "/p/projects/magpie/users/beier/FSECmodeling/output"
)
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{scenarioType}{options: all, or one of the a-e groupings}

\item{save}{binary to save files to output directory}

\item{outputdir}{outputdirectory path}
}
\value{
if file is NULL a ggplot2 object will be return
}
\description{
creates supplementary plots for FSDP MAgPIE runs for India case study
}
\details{
blub
}
\author{
David M Chen, Vartika Singh
}
