% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplPlotsLandConsv.R
\name{SupplPlotsLandConsv}
\alias{SupplPlotsLandConsv}
\title{SupplPlotsLandConsv}
\usage{
SupplPlotsLandConsv(
  outFolder,
  file = "suppPlotsLandConsv.png",
  gdxFolder = NULL
)
}
\arguments{
\item{outFolder}{output folder}

\item{file}{Name of file output}

\item{gdxFolder}{Run folder}
}
\value{
Map with cell share included in land conservation measure
}
\description{
Creates land conservation map
}
\author{
Patrick v. Jeetze
}
