% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeSupplFood.R
\name{themeSupplFood}
\alias{themeSupplFood}
\title{themeSupplFood}
\usage{
themeSupplFood(base_size = 16)
}
\arguments{
\item{base_size}{base size of plot}
}
\description{
creates supplementary themes for FSDP MAgPIE runs
}
