% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeSupplReg.R
\name{themeSupplReg}
\alias{themeSupplReg}
\title{themeSupplFood}
\usage{
themeSupplReg(
  base_size = 11,
  base_family = "",
  rotate_x = FALSE,
  panel.spacing = 3
)
}
\arguments{
\item{base_size}{plot settings}

\item{base_family}{plot settings}

\item{rotate_x}{plot settings}

\item{panel.spacing}{plot settings}
}
\description{
creates supplementary themes for FSDP MAgPIE runs
}
