% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertReportFSDP.R
\name{convertReportFSDP}
\alias{convertReportFSDP}
\title{convertReportFSDP}
\usage{
convertReportFSDP(rep, subset = FALSE, varlist = NULL)
}
\arguments{
\item{rep}{reporting .rds file (produced by FDSP_collect.R output script)}

\item{subset}{TRUE returns "BAU 2020","BAU 2050","SDP 2050". FALSE returns all scenarios}

\item{varlist}{file name for plain text variable list (e.g. "var_names.csv")}
}
\value{
if file is NULL a ggplot2 object will be return
}
\description{
reads in FSDP reporting file
}
\details{
blub
}
\author{
Florian Humpenoeder
}
