% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariables}
\alias{getVariables}
\title{getVariables}
\usage{
getVariables(reportVariables = NULL)
}
\arguments{
\item{reportVariables}{Variables in report file}
}
\value{
named vector
}
\description{
Variables for FSDP plotting
}
\details{
includes information about unit, variable group, rounding etc
}
\author{
Florian Humpenoeder
}
