% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapFSDP.R
\name{heatmapFSDP}
\alias{heatmapFSDP}
\title{heatmapFSDP}
\usage{
heatmapFSDP(repReg, regionSel = "GLO", file = NULL)
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with merge_report.R output script. To produce a figure with a reduced size, merge only the scenarios you want to have in the plot, e.g. using the filter command when doing the runselection, and filtering for FSEC[ab] for the scenarios in table 1.}

\item{regionSel}{Region that should be plotted}

\item{file}{file name (e.g. FSDP_heatmap.pdf or FSDP_heatmap.jpg) or NULL}
}
\value{
if file is NULL a ggplot2 object will be return
}
\description{
creates a heatmap for FSDP MAgPIE runs
}
\details{
blub
}
\author{
Florian Humpenoeder
}
