% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapFSDP.R
\name{heatmapFSDP}
\alias{heatmapFSDP}
\title{heatmapFSDP}
\usage{
heatmapFSDP(repReg, regionSel = "GLO", tableType = 1, file = NULL)
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{regionSel}{Region that should be plotted. select "IND" for India plots}

\item{tableType}{options: 1 (FSECa,FSECc), 2 (FSECb,FSECc,FSECd)}

\item{file}{file name (e.g. FSDP_heatmap.pdf or FSDP_heatmap.jpg) or NULL}
}
\value{
if file is NULL a ggplot2 object will be return
}
\description{
creates a heatmap for FSDP MAgPIE runs
}
\details{
blub
}
\author{
Florian Humpenoeder, Vartika Singh
}
