% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateHealthImpacts.R
\name{integrateHealthImpacts}
\alias{integrateHealthImpacts}
\title{integrateHealthImpacts}
\usage{
integrateHealthImpacts(healthImpacts_filePath, scenario, dir = ".")
}
\arguments{
\item{healthImpacts_filePath}{full pathname of the health impacts dataset}

\item{scenario}{name of the scenario}

\item{dir}{scenario output directory}
}
\description{
Reads in the results from Marco Springmann's health impacts model, filters them down to the
FSEC-relevant results, and saves them to each scenarios' respective report.mif.
}
\author{
Michael Crawford
}
