% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplPlotsCropShr.R
\name{SupplPlotsCropShr}
\alias{SupplPlotsCropShr}
\title{SupplPlotsCropShr}
\usage{
SupplPlotsCropShr(
  outFolder,
  file = NULL,
  scenarios = c("BAU", "FSDP"),
  plotyears = c("2020", "2050"),
  panel = "row"
)
}
\arguments{
\item{outFolder}{output folder}

\item{file}{Name of file output}

\item{scenarios}{scenarios used for plotting}

\item{plotyears}{years for the plot}

\item{panel}{plot regions as "row" or as "matrix"}
}
\value{
Crop share on the y-axis and cropland area in each cluster on the x-axis.
}
\description{
Creates crop share plot
}
\details{
blub
}
\author{
Patrick v. Jeetze, Benjamin Bodirsky
}
