% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplPlotsCropShr.R
\name{SupplPlotsCropShr}
\alias{SupplPlotsCropShr}
\title{SupplPlotsCropShr}
\usage{
SupplPlotsCropShr(repGrid, file = NULL, plotyear = "2050", panel = "row")
}
\arguments{
\item{repGrid}{gridded joint fsdp report}

\item{file}{Name of file output}

\item{plotyear}{year for the plot}

\item{panel}{plot regions as "row" or as "matrix"}
}
\value{
Crop share on the y-axis and cropland area in each cluster on the x-axis.
}
\description{
Creates crop share plot
}
\details{
blub
}
\author{
Patrick v. Jeetze, Benjamin Bodirsky
}
