% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplPlotsCropShr.R
\name{SupplPlotsCropShr}
\alias{SupplPlotsCropShr}
\title{SupplPlotsCropShr}
\usage{
SupplPlotsCropShr(
  gdx,
  folder = "",
  file = NULL,
  plotyear = "y2050",
  panel = "row"
)
}
\arguments{
\item{gdx}{gdx File}

\item{folder}{output folder ending with a slash}

\item{file}{Name of file output}

\item{panel}{plot regions as "row" or as "matrix"}

\item{year}{year to be plotted}
}
\value{
Crop share on the y-axis and cropland area in each cluster on the x-axis.
}
\description{
Creates crop share plot
}
\details{
blub
}
\author{
Patrick v. Jeetze, Benjamin Bodirsky
}
