% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplPlots.R
\name{SupplPlotsFSDP}
\alias{SupplPlotsFSDP}
\title{SupplPlotsFSDP}
\usage{
SupplPlotsFSDP(
  repReg,
  scenarioType = "all",
  save = TRUE,
  outputdir = "/p/projects/magpie/users/beier/FSECmodeling/output"
)
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{scenarioType}{options: all, or one of the a-e groupings}

\item{save}{binary to save files to output director}

\item{outputdir}{outputdirectory path}
}
\value{
if file is NULL a ggplot2 object will be return
}
\description{
creates supplementary plots for FSDP MAgPIE runs
}
\details{
blub
}
\author{
David M Chen
}
