% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplPlotsFSDP.R
\name{SupplPlotsFSDP}
\alias{SupplPlotsFSDP}
\title{SupplPlotsFSDP}
\usage{
SupplPlotsFSDP(
  repReg,
  scenarioType = "manuscript",
  outFolder,
  calorieSupply = TRUE,
  caseRegion = NULL
)
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{scenarioType}{options: all, or one of the a-e groupings}

\item{outFolder}{output folder to save to}

\item{calorieSupply}{parameter to set if calorie figures needed for calories supply or calorie intake, set FALSE for Intake}

\item{caseRegion}{set to a particular region (i.e. "IND")
if plots are being used for country case studies}
}
\value{
if file is NULL a ggplot2 object will be return
}
\description{
creates supplementary plots for FSDP MAgPIE runs
}
\author{
David M Chen, Vartika Singh
}
