% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationFSDP.R
\name{validationFSDP}
\alias{validationFSDP}
\title{validationFSDP}
\usage{
validationFSDP(repReg, valfile, folder = "output")
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{valfile}{validation file}

\item{folder}{output folder}
}
\description{
Validation for FSDP MAgPIE runs
}
\details{
creates validation for FSDP MAgPIE runs
}
\author{
Florian Humpenoeder
}
