% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationFSDP.R
\name{validationFSDP}
\alias{validationFSDP}
\title{validationFSDP}
\usage{
validationFSDP(
  repReg,
  val,
  regionSel = "aggregate",
  folder = "output",
  scens = "BAU_FSEC",
  showHistLegend = TRUE
)
}
\arguments{
\item{repReg}{rds file or data.frame with all MAgPIE runs, produced with FSDP_collect.R output script.}

\item{val}{rds file or data.frame with validation data}

\item{regionSel}{Region that should be plotted (e.g. c("IND","EUR","GLO")). Aggregate will return LIR, MIR and HIR.}

\item{folder}{output folder}

\item{scens}{if "BAU_FSEC", BAU and FSEC scenarios are plotted, with "central" plots the core scenarios, "extended" plot the core scenarios and all ssps.}

\item{showHistLegend}{show legend for historic data sets. TRUE / FALSE}
}
\description{
Validation for FSDP MAgPIE runs
}
\details{
creates validation for FSDP MAgPIE runs
}
\author{
Florian Humpenoeder
}
